small debug changes

+++ ServerUtility/WebCacher.cpp
@@ -2,14 +2,14 @@
 #include "WebCacher.hpp"
 
 WebCacher::WebCacher(std::function<std::string()>  generator_f,
-	int ttl_millis):ttl(ttl_millis),generator(generator_f){
+	long ttl_millis):ttl(ttl_millis),generator(generator_f){
 	lastUpdate = 0;
 }
 
 std::string WebCacher::getResponse(){
     std::lock_guard<std::mutex> guard(response_mutex);
 
-	int millis = TimeUtility::getCurrentMillis();
+	long millis = TimeUtility::getCurrentMillis();
 	if (millis - lastUpdate > ttl){
 		response = generator();
 		lastUpdate = TimeUtility::getCurrentMillis();
